requirements.txt
numpy==2.1.3
Pillow==10.0.1
Pillow==11.0.0
Requests==2.32.3
streamlit==1.40.1
tensorflow==2.14.0
tensorflow_intel==2.14.0
